+++ client/projects/cms/src/app/modules/dashboard/modules/module-instance/interfaces/single-service.interface.ts
@@ -1,0 +1,6 @@
+import {Observable} from 'rxjs';
+
+export interface SingleService {
+	get(moduleId: string, id: string): Observable<any>;
+	save(moduleId: string, id: string, data: any, metadata?: any): Observable<any>;
+}