+++ client/projects/cms/src/app/modules/dashboard/modules/module-instance/utils/generate-filters.ts
@@ -1,0 +1,16 @@
+import {FilterMethod, Module} from 'definitions';
+import {WhereFilter} from '../../../../../shared/interfaces/where-filter.interface';
+
+export function generateFilters(
+	module: Module,
+	search: string,
+	filter: WhereFilter[]
+) {
+	return search && search.trim() ?
+		[{
+			key: module.layout.searchModule.key,
+			operator: module.layout.searchModule.simple ? FilterMethod.Equal : FilterMethod.ArrayContains,
+			value: search.trim().toLowerCase()
+		}] :
+		filter;
+}