+++ client/projects/cms/src/app/shared/modules/eval/eval.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { EvalPipe } from './eval.pipe';
+
+describe('EvalPipe', () => {
+  it('create an instance', () => {
+    const pipe = new EvalPipe();
+    expect(pipe).toBeTruthy();
+  });
+});