+++ client/projects/shared/blocks/blocks.module.ts
@@ -8,12 +8,18 @@ import {BlogComponent} from './blocks/blog/blog.component';
 import {CommonBlockComponent} from './blocks/common.block';
 import {ContactFormComponent} from './blocks/contact-form/contact-form.component';
 import {ContentComponent} from './blocks/content/content.component';
+import {CtaComponent} from './blocks/email/cta/cta.component';
+import {ParagraphComponent} from './blocks/email/paragraph/paragraph.component';
 import {BlockLinkDirective} from './directives/block-link/block-link.directive';
 
 const B_COMPONENTS = [
   ContentComponent,
   ContactFormComponent,
   BlogComponent,
+
+  // Email
+  ParagraphComponent,
+  CtaComponent
 ];
 
 @NgModule({